pyproject.toml
[tool.poetry]
name = "fabric-o-matic"
version = "0.1.0"
description = ""
authors = ["QwiglyDee <qwiglydee@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fake-bpy-module = "^20240307"

[tool.poetry.group.dev.dependencies]
numpy = "^1.26.3"
scipy = "^1.12.0"
ipykernel = "^6.29.0"
ipywidgets = "^8.1.1"
nbformat = "^5.9.2"
plotly = "^5.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
